refactor: View와 Logic 분리
TodayPage 페이지에 UI 코드와 비즈니스 로직이 섞여있는걸 커스텀 훅으로 분리하여 TodayPage는 화면을 그리는 역할만 하도록 수정

새로운 커스텀 훅인 useTodaySchedules에 API 호출과 데이터 가공, 선택 로직을 이동
